refactor: extract resolveDeployContext helper, deduplicate param parsing across badge and cli
